Update go modules and better error handling

+++ nex/datastore/post_meta_binary.go
@@ -1,21 +1,29 @@
 package nex_datastore
 
 import (
-	"github.com/PretendoNetwork/nex-go"
-	"github.com/PretendoNetwork/nex-protocols-go/datastore"
 	"github.com/PretendoNetwork/team-kirby-clash-deluxe-secure/database"
 	"github.com/PretendoNetwork/team-kirby-clash-deluxe-secure/globals"
+
+	"github.com/PretendoNetwork/nex-go"
+	datastore "github.com/PretendoNetwork/nex-protocols-go/datastore"
+	datastore_types "github.com/PretendoNetwork/nex-protocols-go/datastore/types"
 )
 
-func PostMetaBinary(err error, client *nex.Client, callID uint32, param *datastore.DataStorePreparePostParam) {
+func PostMetaBinary(err error, client *nex.Client, callID uint32, param *datastore_types.DataStorePreparePostParam) uint32 {
+	if err != nil {
+		globals.Logger.Error(err.Error())
+		return nex.Errors.DataStore.InvalidArgument
+	}
+
 	metaBinary := database.GetMetaBinaryByOwnerPID(client.PID())
 
-	if metaBinary.DataID != 0 {
-		// * Meta binary already exists
-		if param.PersistenceInitParam.DeleteLastObject {
-			// * Delete existing object before uploading new one
-			// TODO - Check error
-			_ = database.DeleteMetaBinaryByDataID(metaBinary.DataID)
+	// * Meta binary already exists
+	if metaBinary.DataID != 0 && param.PersistenceInitParam.DeleteLastObject {
+		// * Delete existing object before uploading new one
+		err = database.DeleteMetaBinaryByDataID(metaBinary.DataID)
+		if err != nil {
+			globals.Logger.Critical(err.Error())
+			return nex.Errors.DataStore.Unknown
 		}
 	}
 
@@ -44,4 +52,6 @@ func PostMetaBinary(err error, client *nex.Client, callID uint32, param *datasto
 	responsePacket.AddFlag(nex.FlagReliable)
 
 	globals.NEXServer.Send(responsePacket)
+
+	return 0
 }
